Reuse volfile generation and notification code
...in volume create and volume set.
